+++ src/IdentityOverEqualityTrait.php
@@ -85,64 +85,6 @@ if (version_compare(PHPUnitVersionRetriever::getVersion(), '7.0.0') < 0) {
 } else {
     trait IdentityOverEqualityTrait
     {
-        public static function assertEquals($expected, $actual, string $message = '', float $delta = 0.0, int $maxDepth = 10, bool $canonicalize = false, bool $ignoreCase = false): void
-        {
-            $trace = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS, 2);
-
-            // sometimes, PHPUnit calls `assertEquals` instead of `assertSame` internally, we allow that
-            if (!in_array(
-                $trace[1]['function'],
-                [
-                    'assertJsonStringEqualsJsonString',
-                    'assertSame',
-                    'assertXmlStringEqualsXmlString',
-                ],
-                true
-            )) {
-                Reporter::report('Use `->assertSame()` instead of `->assertEquals()`.');
-            }
-
-            call_user_func_array(['parent', __FUNCTION__], func_get_args());
-        }
-
-        public static function assertNotEquals($expected, $actual, string $message = '', $delta = 0, $maxDepth = 10, $canonicalize = false, $ignoreCase = false): void
-        {
-            $trace = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS, 2);
-
-            // sometimes, PHPUnit calls `assertEquals` instead of `assertSame` internally, we allow that
-            if ('assertNotSame' !== $trace[1]['function']) {
-                Reporter::report('Use `->assertNotSame()` instead of `->assertNotEquals()`.');
-            }
-
-            call_user_func_array(['parent', __FUNCTION__], func_get_args());
-        }
-
-        public static function assertAttributeEquals($expected, string $actualAttributeName, $actualClassOrObject, string $message = '', float $delta = 0, int $maxDepth = 10, bool $canonicalize = false, bool $ignoreCase = false): void
-        {
-            // need to override the method, as original on v4 is not using Late Static Binding
-            static::assertEquals(
-                $expected,
-                self::readAttribute($actualClassOrObject, $actualAttributeName),
-                $message,
-                $delta,
-                $maxDepth,
-                $canonicalize,
-                $ignoreCase
-            );
-        }
-
-        public static function assertAttributeNotEquals($expected, string $actualAttributeName, $actualClassOrObject, string $message = '', float $delta = 0, int $maxDepth = 10, bool $canonicalize = false, bool $ignoreCase = false): void
-        {
-            // need to override the method, as original on v4 is not using Late Static Binding
-            static::assertNotEquals(
-                $expected,
-                self::readAttribute($actualClassOrObject, $actualAttributeName),
-                $message,
-                $delta,
-                $maxDepth,
-                $canonicalize,
-                $ignoreCase
-            );
-        }
+        use IdentityOverEqualityTrait7;
     }
 }
